requirements.txt
requests
pandas
python-dotenv
MetaTrader5
discord
streamlit